Filter streets based on difficulty in-memory rather than in-query
This fixes an issue introduced with the diffculty setting; imagine if a street was broken into multiple elements, some of which were only usable for a higher difficulty, and you selected the tourist setting. The overpass API would only return some of the street. We join all elements/segments in-memory because of similar issues like this we found in the past, but this time we wouldn't even have the other segments. So now we ask for all streets and filter on the client-side.